requirements.txt
# 学思通核心依赖
cryptography>=41.0.0      # 加密库
qrcode>=7.4.0             # 二维码生成
Pillow>=10.0.0            # 图片处理（勋章生成）
